Moved gamemode list back into gamemode handler
